Call set_buffer_uptodate() in blockget()
The uptodate rule in userspace is different from kernel. In kernel, if
buffer is not uptodate, it will try to read from disk.

So, if we did the following, the data will lost.

    buffer = blockget(10)
    <set data to buffer>
    mark_buffer_dirty(buffer)
    buffer = blockread(10) <- read from disk

In blockget(), this calls set_buffer_uptodate() before return it,
without any initialization. The caller have to initialize data.

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>
Signed-off-by: Daniel Phillips <daniel@tux3.org>