Introduce mark_btree_dirty() as example of btree_inode()
This fixes forgeting mark_inode_dirty() after btree root change as
example of btree_inode(). This will be replaced with atomic commit
though.

Another example is, we can know btree->sb from btree_inode()->i_sb.
It would be the example of itable was moved to itable to volmap->btree.

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>
Signed-off-by: Daniel Phillips <daniel@tux3.org>