Change bufferdirty() strategy
To make stable buffer for backend, this forks buffers instead of only
data.  With this, backend will become simpler. Because the backend
doesn't need to care about block-fork, when it is flushing buffers.

To do it, this return forked buffer.  [The error check is FIXME for
now, it will be fixed with kernel change. Because blockdirty() for
kernel, it will return -EAGAIN as special case.]

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>
Signed-off-by: Daniel Phillips <daniel@tux3.org>