Preparation for writeback flush on userspace
This prepares to introduce writeback flush stuff to userspace. For it,
this adds writeback.c. The writeback.c will be used to provide
writeback stuff, and will also resolve the dependency of source codes.

With this, buffer.c provides lower level functions only, and
mark_*_dirty() stuff uses those in writeback.c.

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>
Signed-off-by: Daniel Phillips <daniel@tux3.org>