Fix freeing block in btree_chop()
When freeing block, we have to clean dirty buffer on replay too. To do
it, this adds new LOG_LEAF_FREE and LOG_BNODE_FREE to tell it
(LOG_BNODE_MERGE imply it too).

In btree_chop(), we don't need to defer bfree, because we know the
block is redirected, so, we can just call bfree() directly.

Unlikely btree_chop(), in free_empty_btree case, we checks
buffer_dirty() to know if we can use bfree() directly or not.

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>