Fix variable type confusion
Now, we are using uint64_t for u64. But, uint64_t can be "unsigned
long long" or "unsigned long". It bothers us when using printf().

Luckly, in kernel, all arches uses "unsigned long long" nowaday. So,
this changes u64 to "unsigned long long" instead of uint64_t, cleanup
other types too. With this change, we can avoid (L) cast.

For the block_t, it is for sector_t in kernel, and loff_t can be
"long" or "long long" in userland (not printf friendly). And sector_t
can be "unsigned long" or "u64". (32bits arch 32bits is too small).

So, for now, it is "signed long long".

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>