Fix buffer_new buffer handling in write_begin()
If marked buffer as new, block library does some job for it. But,
basically it is implemented based on non delayed-write FS.

And also like write_end(), we have to replace mark_buffer_dirty() by
tux3_mark_buffer_dirty().

This fix buffer_new handling and copy block_write_begin(),

1) Clear buffer_new in blockget().
Otherwise, block library might be confused by it. Caller of blockget()
cares about all of buffer handling (e.g. initialize buffer data,
marking dirty, etc.). So this just clear buffer_new.

2) Remove unmap_underlying_metadata() in write_begin()
We don't use blockdev buffer at all. So we don't need this at all.

3) Replace mark_buffer_dirty() by tux3_mark_buffer_dirty() in write_begin()
If get_block() marked buffer as new, it thinks new block was allocated
(this is not true if delayed allocation). And If page is uptodate, it
clear new and mark buffer dirty, because if buffer is still new on error,
page_zero_new_buffers() will fill buffer by zero. It is data loss.

But I'm not sure we need to mark buffer dirty in this case. When it
was hole, we will not need to mark buffer dirty, because tux3 (delayed
allocation) is not allocated block yet.

And if mmaped page and have data, buffer should already be dirty?

Well, so add "FIXME:" as comment there. And we should revisit there
after mmap support.

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>