Fix "Fix change_begin/end chunk for write(2) family"
With the previous change, change_begin/end was separated to small
transactions. And it may change only data pages.

I.e.

before change:
    mtime and whole data

after change:
    mtime and small chunk, then small chunk, small chunk, ...

So, inode attrs are may not be dirty on some delta. The problem is
here.

For a regular file, and even if iattrs are clean, we have to provide
stable idata for backend.

Because backend may be committing data pages. If so, backend have to
check idata->i_size, and may save dtree root. But previous delta
doesn't have stable iattrs.

So, this provides stable iattrs for regular file, even if previous
delta is clean.

Other types don't have this problem, because:
- Never dirty iattr (e.g. volmap). IOW, iattrs are always stable.
- Or dirty iattr with data, e.g. directory updates timestamp too with
  data blocks.

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>