Introduce tux3_under_backend()
This provides the information to know whether running under
backend. This is useful to test access to backend objects (no lock,
but only backend can access).

To do it, this uses ->journal_info like change_begin/end. The backend
is exclusive with change_begin/end.

Note, this can be used only on process context.

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>