Optimize lookup for deferred allocation inums
For now, we are looking up the list to find deferred allocation
inum. It is slow, of course.

If backend didn't apply deferred inodes to itable, the list can be so
long, and it became the cause of bottleneck.

So, this looks up deferred inums by inode_hash, instead of the
list. However, this is still inefficient, and need to take global
locks.

FIXME: we should use better way something like bitmap, and cooperative
with itable.

Signed-off-by: OGAWA Hirofumi <hirofumi@mail.parknet.co.jp>